refactor(evaluatorq-py): move retry helper to common/ (RES-897)
