+++ packages/evaluatorq-py/src/evaluatorq/simulation/generators/first_message_generator.py
@@ -21,7 +21,7 @@ from evaluatorq.simulation.utils.prompt_builders import (
     build_persona_system_prompt,
     build_scenario_user_context,
 )
-from evaluatorq.simulation.utils.retry import with_retry
+from evaluatorq.common.retry import with_retry
 
 logger = logging.getLogger(__name__)
